Moved main method to other class
