+++ src/main/java/org/roda/rodain/rules/ui/RuleModalProcessing.java
@@ -16,8 +16,7 @@ import javafx.scene.layout.HBox;
 import javafx.scene.layout.VBox;
 
 import org.roda.rodain.core.AppProperties;
-import org.roda.rodain.core.Main;
-import org.roda.rodain.inspection.InspectionPane;
+import org.roda.rodain.core.RodaIn;
 import org.roda.rodain.rules.Rule;
 import org.roda.rodain.rules.VisitorStack;
 import org.roda.rodain.rules.VisitorState;
@@ -151,6 +150,6 @@ public class RuleModalProcessing extends BorderPane {
   private void close() {
     timer.cancel();
     RuleModalController.cancel();
-    Main.getInspectionPane().updateRuleList();
+    RodaIn.getInspectionPane().updateRuleList();
   }
 }
